Reorganize composables
- Move all composables to lib/composables
- Indicate private composables by "_" prefix
- Remove deprecated KResponsiveWindow mixin documentation page
  in favor of a new useKResponsiveWindow page
