+++ lib/composables/_useKWindowDimensions/__tests__/index.spec.js
@@ -1,6 +1,6 @@
 import { defineComponent } from '@vue/composition-api';
 import { mount } from '@vue/test-utils';
-import useKWindowDimensions from '../useKWindowDimensions';
+import useKWindowDimensions from '../';
 
 const resizeWindow = (width, height) => {
   window.innerWidth = width;
