+++ lib/composables/_useKWindowDimensions/index.js
@@ -1,4 +1,4 @@
-import './composition-api'; //Due to @vue/composition-api shortcomings, add plugin prior to use in kolibri, studio and tests
+import '../composition-api'; //Due to @vue/composition-api shortcomings, add plugin prior to use in kolibri, studio and tests
 import { onMounted, onUnmounted, ref } from '@vue/composition-api';
 import { throttle } from 'frame-throttle';
