+++ lib/composables/useKResponsiveWindow/__tests__/index.spec.js
@@ -2,7 +2,7 @@ import 'mock-match-media/jest-setup.cjs';
 import { setMedia } from 'mock-match-media';
 import { defineComponent } from '@vue/composition-api';
 import { mount } from '@vue/test-utils';
-import useKResponsiveWindow from '..';
+import useKResponsiveWindow from '../';
 
 const resizeWindow = (width, height = 768) => {
   window.innerWidth = width;
