+++ lib/composables/useKResponsiveWindow/index.js
@@ -1,5 +1,5 @@
 import { computed, onBeforeUnmount, onMounted, ref, watch } from '@vue/composition-api';
-import useKWindowDimensions, { windowWidth, windowHeight } from '../useKWindowDimensions';
+import useKWindowDimensions, { windowWidth, windowHeight } from '../_useKWindowDimensions';
 import MediaQuery from './MediaQuery';
 
 /** Global variables */
@@ -29,9 +29,20 @@ const heightQuery = new MediaQuery('screen and (max-height: 600px)', event => {
 });
 
 /**
+ * Check if Nuxt is server side rendering
+ * @returns {Boolean}
+ */
+function isNuxtServerSideRendering() {
+  return process && process.server;
+}
+
+/**
  * Initialize media query window properties
  */
 function initProps() {
+  if (isNuxtServerSideRendering) {
+    return;
+  }
   if (window.matchMedia) {
     orientationQuery.eventHandler(orientationQuery.mediaQueryList);
     heightQuery.eventHandler(heightQuery.mediaQueryList);
