+++ lib/composables/useKShow/__tests__/index.spec.js
@@ -1,4 +1,4 @@
-import useKShow from '../index';
+import useKShow from '../';
 
 const { show } = useKShow();
